Update to 1.9.11-git-beta1-r3
Redundancy 1.9.11 Changelog
===========================
Bugfixes
========
- Fixed a bug when copy- or moving a file when a image was missing
- Fixed a bug producing wrong filesystem stats when a extension is
lower(".png") or uppercase ".PNG". The case is now ignored
- Fixed a bug when a login was stopped because of XSS content a wrong
input will be executed in the database
- Fixed the possibility of a wrong calculation of a size in
getDirectorySize()
- Changed a column type in the database to allow sorting by date
- Fixed a bug causing that the contingent could not be set
- Fixed a bug loosing the date at moving a file or dir
- Fixed a bug which caused wrong set of the timestamps
- Fixed a bug which caused no redirection after logout
- The user can't paste the source folder into itself anymore from web
interface
- 58 methods where improved
Security Bugfixes
=================
- Added a security question when deleting an account from the
administration panel
- Fixed a security vulnerability which allowed under some cases to see
the api token while be logged in as a guest
- Added an .htaccess file for the Includes directory
- Fixed a bug in https redirecting causing a warning on the server
- Fixed a sql injection leak in several methods
- Added Program_Session_Timeout to set a timeout for killing a too old
session
Feature additions/ Improvements
===============================
- Added a display correction if the extension is uppercase
- Added an error message when a file is not found
- Added an error message when a dir is not found
- Added the function to set a new api token
- The administration panel is now bootstrap powered
- Program_Enable_Mail added to control mail support
- Program_Enable_Banning to control banning
- New API added (still work in progress! Not recommended to use yet)
- Program_Enable_Chart added to control drawing of the chart
- Added missing translations at the admin panel
- Added Program_Branding to control the path of the branding logo
- Added a message after deleting a filesystem entry
- Replaced the deprecated call to get the mimetype in download.inc.php
with the regular call from the kernel
- In the mobile view, the current page is selected in the menu
- Added Program_Enable_ErrorHandler to control internal error handler
which logs into System.log
- The error handler logs internal warnings and errors
- Added an error message for the changes module when no changes could be
determined
Features removed
================
- User_Activation_Link removed
- User_Activation_Link_Sender removed
- User_Activation_Link_Sender_Email removed
- Removed Program_Embed_GPL_Header

+++ Includes/goodbye.inc.php
@@ -25,35 +25,38 @@
 	//Include uri check
 	require_once ("checkuri.inc.php");
 	//Use with caution
-	//TODO: Find the bug which causes a complete loss of any data?! :(
 	if (isset($_SESSION) == false)
 			session_start();
-	if ($_SESSION["role"] != 3 && isset($_GET["sure"])){	
+	if ($_SESSION["role"] != 3 && isset($_GET["sure"]) && isGuest() == false){	
 		include $GLOBALS["Program_Dir"]."Includes/DataBase.inc.php";	
 		$userID = mysqli_real_escape_string($connect,$_SESSION["user_id"]);	
-		$getFiles = mysqli_query($connect,"Select * from Files where UserID = '$userID'") ;
-		while ($row = mysqli_fetch_object($getFiles)) {
-			echo "<br>Deleting ".$row->Displayname ."...";
+		$getFiles = mysqli_query($connect,"Select Displayname,Filename from Files where UserID = '$userID'") ;
+		while ($row = mysqli_fetch_object($getFiles)) {			
 			if ($row->Filename != $row->Displayname)
-			unlink($GLOBALS["Program_Dir"]."Storage/".$row->Filename);
-			echo "<br>Removing database entry...";
+				unlink($GLOBALS["Program_Dir"]."Storage/".$row->Filename);					
 			mysqli_query($connect,"Delete from Files where UserID = '$userID'");			
-			mysqli_query($connect,"Delete from Share where UserID = '$userID'");
-			echo "<br>Removing shares entries ...";
-			echo "..Done";			
+			mysqli_query($connect,"Delete from Share where UserID = '$userID'");						
 		}		
 		mysqli_query($connect,"Delete from Users where ID = '$userID'");
-		mysqli_close($connect);	
-		if (isset($_SESSION) == false)
-			session_start();
-			//exit everything
-			session_unset();
-			session_destroy();
+		mysqli_close($connect);			
+		//exit everything
+		session_unset();
+		session_destroy();
+		header("Location: index.php");
+		exit;
 	}	
 	else
 	{
-		echo $GLOBALS["Program_Language"]["Delete_Warning"]."<br><br>";
-		echo "<a href = 'index.php?module=goodbye&sure=true'>".$GLOBALS["Program_Language"]["Delete_OK"]."</a>";
+		?>
+			<h2>
+				<?php echo $GLOBALS["Program_Language"]["Delete_Warning"];?>
+			</h2>
+			<br>
+			<br>
+			<a href = 'index.php?module=goodbye&sure=true'>
+				<? echo$GLOBALS["Program_Language"]["Delete_OK"];?>			
+			</a>";	
+		<?php
 	}
 ?>
 </div>
